# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/collections": {
      "get": {
        "operationId": "CollectionController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limit",
            "schema": {
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Offset",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search by name substring",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionsListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Collections"
        ]
      },
      "post": {
        "operationId": "CollectionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionDto"
                }
              }
            }
          }
        },
        "tags": [
          "Collections"
        ]
      }
    },
    "/api/collections/{id}": {
      "get": {
        "operationId": "CollectionController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionDto"
                }
              }
            }
          }
        },
        "tags": [
          "Collections"
        ]
      },
      "patch": {
        "operationId": "CollectionController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionDto"
                }
              }
            }
          }
        },
        "tags": [
          "Collections"
        ]
      },
      "delete": {
        "operationId": "CollectionController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Collections"
        ]
      }
    },
    "/api/collections/{id}/documents": {
      "get": {
        "operationId": "CollectionController_documents",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limit",
            "schema": {
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Offset",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Collections"
        ]
      }
    },
    "/api/documents": {
      "get": {
        "operationId": "DocumentsController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limit",
            "schema": {
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Offset",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "title",
            "required": false,
            "in": "query",
            "description": "Filter by title (substring match)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Full-text query applied to title OR content (substring match)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scraperId",
            "required": false,
            "in": "query",
            "description": "Filter by scraper id (UUID)",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "description": "Date from (inclusive)",
            "schema": {
              "format": "date-time",
              "example": "2024-01-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "description": "Date to (inclusive)",
            "schema": {
              "format": "date-time",
              "example": "2024-12-31T23:59:59.999Z",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List documents with pagination",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentsListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Documents"
        ]
      },
      "post": {
        "operationId": "DocumentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created document",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDto"
                }
              }
            }
          }
        },
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/documents/{id}": {
      "get": {
        "operationId": "DocumentsController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDto"
                }
              }
            }
          }
        },
        "tags": [
          "Documents"
        ]
      },
      "patch": {
        "operationId": "DocumentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated document",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDto"
                }
              }
            }
          }
        },
        "tags": [
          "Documents"
        ]
      },
      "delete": {
        "operationId": "DocumentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete result",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Documents"
        ]
      }
    },
...
  },
  "info": {
    "title": "Trade API",
    "description": "Auto-generated OpenAPI documentation for Trade backend",
    "version": "1.0.0",
    "contact": {
      "name": "API",
      "url": "",
      "email": ""
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "schemas": {
...
      "CollectionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Collection UUID"
          },
          "name": {
            "type": "string",
            "description": "Collection name"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "filters": {
            "type": "object",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "description": "Created at (ISO 8601)"
          },
          "updated_at": {
            "type": "string",
            "description": "Updated at (ISO 8601)"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "filters",
          "created_at",
          "updated_at"
        ]
      },
      "CollectionsListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CollectionDto"
            }
          },
          "total": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
      "CollectionCreateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Collection name"
          },
          "description": {
            "type": "object",
            "description": "Description"
          },
          "filters": {
            "type": "object",
            "description": "Filter criteria"
          }
        },
        "required": [
          "name"
        ]
      },
      "CollectionUpdateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Collection name"
          },
          "description": {
            "type": "object",
            "description": "Description"
          },
          "filters": {
            "type": "object",
            "description": "Filter criteria"
          }
        }
      },
      "DocumentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "scraper": {
            "$ref": "#/components/schemas/ScraperShortDto"
          },
          "scrapedAt": {
            "type": "string",
            "format": "date-time"
          },
          "contentHash": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "scraper",
          "scrapedAt",
          "contentHash"
        ]
      },
      "DocumentsListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentDto"
            }
          },
          "total": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
      "ScraperShortDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "DocumentCreateDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "scraperId": {
            "type": "string",
            "description": "UUID of scraper",
            "format": "uuid"
          },
          "content": {
            "type": "object",
            "description": "Content of the document. If non-string is provided will be stringified."
          },
          "scrapedAt": {
            "type": "string",
            "description": "Scraped at date",
            "format": "date-time"
          },
          "contentHash": {
            "type": "string",
            "description": "Content hash. If omitted, server computes sha256(content)."
          }
        },
        "required": [
          "title",
          "scraperId",
          "content"
        ]
      },
      "DocumentUpdateDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "object",
            "description": "Content of the document. If non-string is provided will be stringified."
          },
          "scraperId": {
            "type": "string",
            "description": "UUID of scraper",
            "format": "uuid"
          },
          "scrapedAt": {
            "type": "string",
            "description": "Scraped at date",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}